\name{model.matrix.pffr}
\alias{model.matrix.pffr}
\title{Obtain model matrix for a pffr fit}
\usage{
  \method{model.matrix}{pffr} (object, ...)
}
\arguments{
  \item{object}{a fitted \code{pffr}-object}

  \item{...}{other arguments, passed to
  \code{\link[mgcv]{predict.gam}}.}
}
\value{
  A model matrix
}
\description{
  Obtain model matrix for a pffr fit
}
\author{
  Fabian Scheipl
}

